+++ examples/esp32-c3-example/src/main.rs
@@ -2,20 +2,24 @@
 #![no_main]
 
 use embassy_executor::Spawner;
+use embassy_time::Duration;
 use esp_backtrace as _;
 use esp_hal::{
-    interrupt::software::SoftwareInterruptControl,
     rng::{Trng, TrngSource},
     timer::timg::TimerGroup,
 };
-use esp_println::println;
 use esp_radio::ble::controller::BleConnector;
 use esp_xbox_controller::{ControllerConfig, ControllerSelector, run};
+use log::{info, warn};
 use trouble_host::prelude::ExternalController;
 
 esp_bootloader_esp_idf::esp_app_desc!();
 
-const CONTROLLER: ControllerConfig = ControllerConfig::new(ControllerSelector::AnyXbox, None);
+const CONTROLLER: ControllerConfig = ControllerConfig::new(
+    ControllerSelector::AnyXbox,
+    None,
+    Duration::from_millis(10000),
+);
 
 #[esp_rtos::main]
 async fn main(_spawner: Spawner) -> ! {
@@ -24,8 +28,7 @@ async fn main(_spawner: Spawner) -> ! {
     esp_alloc::heap_allocator!(size: 72 * 1024);
 
     let timer_group = TimerGroup::new(peripherals.TIMG0);
-    let software_interrupts = SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
-    esp_rtos::start(timer_group.timer0, software_interrupts.software_interrupt0);
+    esp_rtos::start(timer_group.timer0, peripherals.FROM_CPU_INTR0);
 
     let _trng_source = TrngSource::new(peripherals.RNG, peripherals.ADC1);
     let mut random = Trng::try_new().expect("failed to initialize true random generator");
@@ -38,8 +41,8 @@ async fn main(_spawner: Spawner) -> ! {
         &mut random,
         CONTROLLER,
         |notification| match notification {
-            Ok(state) => println!("controller state: {:?}", state),
-            Err(error) => println!("input report parse error: {:?}", error),
+            Ok(state) => info!("controller state: {:?}", state),
+            Err(error) => warn!("input report parse error: {:?}", error),
         },
     )
     .await